reuse cidr flag completion from ssh cmd in ssh-patch
