+++ pkg/cmd/base/options.go
@@ -20,16 +20,14 @@ import (
 
 //go:generate mockgen -destination=./mocks/mock_options.go -package=mocks github.com/gardener/gardenctl-v2/pkg/cmd/base CommandOptions
 
-// CommandOptions is the base interface for command options.
-type CommandOptions interface {
+// Runnable is the base interface for command options.
+type Runnable interface {
 	// Complete adapts from the command line args to the data required.
 	Complete(util.Factory, *cobra.Command, []string) error
 	// Validate validates the provided options.
 	Validate() error
 	// Run does the actual work of the command.
 	Run(util.Factory) error
-	// AddFlags adds flags to adjust the output to a cobra command.
-	AddFlags(*pflag.FlagSet)
 }
 
 // Options contains all settings that are used across all commands in gardenctl.
@@ -41,10 +39,10 @@ type Options struct {
 	Output string
 }
 
-var _ CommandOptions = &Options{}
+var _ Runnable = &Options{}
 
 // WrapRunE creates a cobra RunE function that has access to the factory.
-func WrapRunE(o CommandOptions, f util.Factory) func(cmd *cobra.Command, args []string) error {
+func WrapRunE(o Runnable, f util.Factory) func(cmd *cobra.Command, args []string) error {
 	return func(cmd *cobra.Command, args []string) error {
 		if err := o.Complete(f, cmd, args); err != nil {
 			return fmt.Errorf("failed to complete command options: %w", err)
