+++ pkg/cmd/ssh/access_config.go
@@ -1,3 +1,9 @@
+/*
+SPDX-FileCopyrightText: 2022 SAP SE or an SAP affiliate company and Gardener contributors
+
+SPDX-License-Identifier: Apache-2.0
+*/
+
 package ssh
 
 import (
@@ -9,15 +15,14 @@ import (
 	"time"
 
 	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 
 	"github.com/gardener/gardenctl-v2/internal/util"
-	"github.com/gardener/gardenctl-v2/pkg/cmd/base"
 )
 
-// AccessConfig is a struct used by all ssh related commands.
+// AccessConfig is a struct that is embedded in the options of ssh related commands.
 type AccessConfig struct {
-	base.Options
-
 	// CIDRs is a list of IP address ranges to be allowed for accessing the
 	// created Bastion host. If not given, gardenctl will attempt to
 	// auto-detect the user's IP and allow only it (i.e. use a /32 netmask).
@@ -28,7 +33,7 @@ type AccessConfig struct {
 	AutoDetected bool
 }
 
-func (o *AccessConfig) Complete(f util.Factory, cmd *cobra.Command, args []string) error {
+func (o *AccessConfig) Complete(f util.Factory, cmd *cobra.Command, args []string, ioStreams util.IOStreams) error {
 	if len(o.CIDRs) == 0 {
 		ctx, cancel := context.WithTimeout(f.Context(), 60*time.Second)
 		defer cancel()
@@ -48,7 +53,7 @@ func (o *AccessConfig) Complete(f util.Factory, cmd *cobra.Command, args []strin
 			name = "CIDRs"
 		}
 
-		fmt.Fprintf(o.IOStreams.Out, "Auto-detected your system's %s as %s\n", name, strings.Join(cidrs, ", "))
+		fmt.Fprintf(ioStreams.Out, "Auto-detected your system's %s as %s\n", name, strings.Join(cidrs, ", "))
 
 		o.CIDRs = cidrs
 		o.AutoDetected = true
@@ -69,4 +74,77 @@ func (o *AccessConfig) Validate() error {
 	}
 
 	return nil
+}
+
+func (o *AccessConfig) AddFlags(flags *pflag.FlagSet) {
+	flags.StringArrayVar(&o.CIDRs, "cidr", nil, "CIDRs to allow access to the bastion host; if not given, your system's public IPs (v4 and v6) are auto-detected.")
+}
+
+type (
+	cobraCompletionFunc          func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective)
+	cobraCompletionFuncWithError func(f util.Factory) ([]string, error)
+)
+
+func completionWrapper(f util.Factory, ioStreams util.IOStreams, completer cobraCompletionFuncWithError) cobraCompletionFunc {
+	return func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+		result, err := completer(f)
+		if err != nil {
+			fmt.Fprintf(ioStreams.ErrOut, "%v\n", err)
+			return nil, cobra.ShellCompDirectiveNoFileComp
+		}
+
+		return util.FilterStringsByPrefix(toComplete, result), cobra.ShellCompDirectiveNoFileComp
+	}
+}
+
+func RegisterCompletionFuncsForAccessConfigFlags(cmd *cobra.Command, factory util.Factory, ioStreams util.IOStreams, flags *pflag.FlagSet) {
+	utilruntime.Must(cmd.RegisterFlagCompletionFunc("cidr", completionWrapper(factory, ioStreams, cidrFlagCompletionFunc)))
+}
+
+func cidrFlagCompletionFunc(f util.Factory) ([]string, error) {
+	var addresses []string
+
+	ctx := f.Context()
+
+	publicIPs, err := f.PublicIPs(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	for _, ip := range publicIPs {
+		cidr := ipToCIDR(ip)
+		addresses = append(addresses, fmt.Sprintf("%s\t<public>", cidr))
+	}
+
+	interfaces, err := net.Interfaces()
+	if err != nil {
+		return nil, err
+	}
+
+	includeFlags := net.FlagUp
+	excludeFlags := net.FlagLoopback
+
+	for _, iface := range interfaces {
+		addrs, err := iface.Addrs()
+		if err != nil {
+			return nil, err
+		}
+
+		if is(iface, includeFlags) && isNot(iface, excludeFlags) {
+			for _, addr := range addrs {
+				addressComp := fmt.Sprintf("%s\t%s", addr.String(), iface.Name)
+				addresses = append(addresses, addressComp)
+			}
+		}
+	}
+
+	return addresses, nil
+}
+
+func is(i net.Interface, flags net.Flags) bool {
+	return i.Flags&flags != 0
+}
+
+func isNot(i net.Interface, flags net.Flags) bool {
+	return i.Flags&flags == 0
 }
